backend/requirements.txt
browser-use
fastapi
uvicorn
python-dotenv
playwright